refact(AuthorizedApps): separate components
